+++ content/template/bookingTemplate/participantView.php
@@ -61,9 +61,35 @@
 					$fo->AddItem($f);
 					$contactAttributes = $eduapi->GetAttribute($edutoken, $so->ToString(), $fo->ToString());
 
+					$db = array();
+					/*if($contact->PersonID != 0) {
+						$fo = new XFiltering();
+						$f = new XFilter('PersonID', '=', $contact->PersonID);
+						$fo->AddItem($f);
+						$db = $eduapi->GetPersonAttribute($edutoken, '', $fo->ToString());
+					}*/
+
 					foreach($contactAttributes as $attr)
 					{
-						renderAttribute($attr, false, "contact");
+						$data = null;
+						foreach($db as $d)
+						{
+							if($d->AttributeID == $attr->AttributeID) {
+								switch($d->AttributeTypeID) {
+									case 1:
+										$data = $d->AttributeChecked;
+										break;
+									case 5:
+										$data = $d->AttributeAlternative->AttributeAlternativeID;
+										break;
+									default:
+										$data = $d->AttributeValue;
+										break;
+								}
+								break;
+							}
+						}
+						renderAttribute($attr, false, "contact", $data);
 					}
 
 					?>