+++ includes/_attributeFunctions.php
@@ -1,34 +1,34 @@
 <?php
-function renderAttribute($attribute, $multiple = false, $suffix = "")
+function renderAttribute($attribute, $multiple = false, $suffix = "", $data = null)
 {
 	switch($attribute->AttributeTypeID)
 	{
 		case 1: // Checkbox
-			renderCheckField($attribute, $multiple, $suffix);
+			renderCheckField($attribute, $multiple, $suffix, $data);
 			break;
 		case 2: // Textfält
-			renderTextField($attribute, $multiple, $suffix);
+			renderTextField($attribute, $multiple, $suffix, $data);
 			break;
 		case 3: // Nummerfält
-			renderNumberField($attribute, $multiple, $suffix);
+			renderNumberField($attribute, $multiple, $suffix, $data);
 			break;
 		case 4: // Flervärdesfält
-			//renderTextField($attribute, $multiple, $suffix);
+			//renderTextField($attribute, $multiple, $suffix, $data);
 			break;
 		case 5: // Dropdownlista
-			renderSelectField($attribute, $multiple, $suffix);
+			renderSelectField($attribute, $multiple, $suffix, $data);
 			break;
 		case 6: // Anteckningsfält
-			renderTextAreaField($attribute, $multiple, $suffix);
+			renderTextAreaField($attribute, $multiple, $suffix, $data);
 			break;
 		case 7: // Datumfält
-			//renderDateField($attribute, $multiple, $suffix);
+			//renderDateField($attribute, $multiple, $suffix, $data);
 			break;
 		case 8: // HTML
-			//renderTextAreaField($attribute, $multiple, $suffix);
+			//renderTextAreaField($attribute, $multiple, $suffix, $data);
 			break;
 		case 9: // Checkboxlista
-			//renderCheckboxListField($attribute, $multiple, $suffix);
+			//renderCheckboxListField($attribute, $multiple, $suffix, $data);
 			break;
 		case 10: // Pinkod
 			break;
@@ -38,34 +38,34 @@ function renderAttribute($attribute, $multiple = false, $suffix = "")
 	}
 }
 
-function renderCheckField($attribute, $multiple, $suffix)
+function renderCheckField($attribute, $multiple, $suffix, $data)
 {
 	echo "<label><div class=\"inputLabel\">";
 	echo $attribute->AttributeDescription;
 	echo "</div><div class=\"inputHolder\">";
-	echo "<input type=\"checkbox\" name=\"edu-attr_" . $attribute->AttributeID . ($suffix != "" ? "-" . $suffix : "") . ($multiple ? "[]" : "") . "\" />";
+	echo "<input type=\"checkbox\"" . ($data != null && $data ? " checked=\"checked\"" : "") . " name=\"edu-attr_" . $attribute->AttributeID . ($suffix != "" ? "-" . $suffix : "") . ($multiple ? "[]" : "") . "\" />";
 	echo "</div></label>";
 }
 
-function renderTextField($attribute, $multiple, $suffix)
+function renderTextField($attribute, $multiple, $suffix, $data)
 {
 	echo "<label><div class=\"inputLabel\">";
 	echo $attribute->AttributeDescription;
 	echo "</div><div class=\"inputHolder\">";
-	echo "<input type=\"text\" name=\"edu-attr_" . $attribute->AttributeID . ($suffix != "" ? "-" . $suffix : "") . ($multiple ? "[]" : "") . "\" value=\"" . $attribute->AttributeValue . "\" />";
+	echo "<input type=\"text\" name=\"edu-attr_" . $attribute->AttributeID . ($suffix != "" ? "-" . $suffix : "") . ($multiple ? "[]" : "") . "\" value=\"" . ($data != null ? $data : $attribute->AttributeValue) . "\" />";
 	echo "</div></label>";
 }
 
-function renderNumberField($attribute, $multiple, $suffix)
+function renderNumberField($attribute, $multiple, $suffix, $data)
 {
 	echo "<label><div class=\"inputLabel\">";
 	echo $attribute->AttributeDescription;
 	echo "</div><div class=\"inputHolder\">";
-	echo "<input type=\"number\" name=\"edu-attr_" . $attribute->AttributeID . ($suffix != "" ? "-" . $suffix : "") . ($multiple ? "[]" : "") . "\" value=\"" . $attribute->AttributeValue . "\" />";
+	echo "<input type=\"number\" name=\"edu-attr_" . $attribute->AttributeID . ($suffix != "" ? "-" . $suffix : "") . ($multiple ? "[]" : "") . "\" value=\"" . ($data != null ? $data : $attribute->AttributeValue) . "\" />";
 	echo "</div></label>";
 }
 
-function renderDateField($attribute, $multiple, $suffix)
+function renderDateField($attribute, $multiple, $suffix, $data)
 {
 	echo "<label><div class=\"inputLabel\">";
 	echo $attribute->AttributeDescription;
@@ -74,16 +74,16 @@ function renderDateField($attribute, $multiple, $suffix)
 	echo "</div></label>";
 }
 
-function renderTextAreaField($attribute, $multiple, $suffix)
+function renderTextAreaField($attribute, $multiple, $suffix, $data)
 {
 	echo "<label><div class=\"inputLabel\">";
 	echo $attribute->AttributeDescription;
 	echo "</div><div class=\"inputHolder\">";
-	echo "<textarea name=\"edu-attr_" . $attribute->AttributeID . ($suffix != "" ? "-" . $suffix : "") . ($multiple ? "[]" : "") . "\" rows=\"3\" resizable=\"resizable\">" . $attribute->AttributeValue . "</textarea>";
+	echo "<textarea name=\"edu-attr_" . $attribute->AttributeID . ($suffix != "" ? "-" . $suffix : "") . ($multiple ? "[]" : "") . "\" rows=\"3\" resizable=\"resizable\">" . ($data != null ? $data : $attribute->AttributeValue) . "</textarea>";
 	echo "</div></label>";
 }
 
-function renderSelectField($attribute, $multiple, $suffix)
+function renderSelectField($attribute, $multiple, $suffix, $data)
 {
 	echo "<label><div class=\"inputLabel\">";
 	echo $attribute->AttributeDescription;
@@ -91,13 +91,13 @@ function renderSelectField($attribute, $multiple, $suffix)
 	echo "<select name=\"edu-attr_" . $attribute->AttributeID . ($suffix != "" ? "-" . $suffix : "") . ($multiple ? "[]" : "") . "\">\n";
 	foreach($attribute->AttributeAlternative as $val)
 	{
-		echo "\t<option value=\"" . $val->AttributeAlternativeID . "\">" . $val->AttributeAlternativeDescription . "</option>\n";
+		echo "\t<option" . ($data != null && $data == $val->AttributeAlternativeID ? " selected=\"selected\"" : "") . " value=\"" . $val->AttributeAlternativeID . "\">" . $val->AttributeAlternativeDescription . "</option>\n";
 	}
 	echo "</select>";
 	echo "</div></label>";
 }
 
-function renderCheckboxListField($attribute, $multiple, $suffix)
+function renderCheckboxListField($attribute, $multiple, $suffix, $data)
 {
 	echo "<div class=\"inputLabel\">";
 	echo $attribute->AttributeDescription;